restructured entire app to be MVVM based (still maintaining previous commit features)
